hotfix edição de mesas e retirada de jogadores

+++ views/mesas/edit.php
@@ -5,6 +5,7 @@ if (!isset($_SESSION['user_id'])) {
     exit();
 }
 require '../../controllers/MesaController.php';
+require '../../includes/db.php';
 
 $mesaController = new MesaController($pdo);
 $mesa_id = $_GET['id'] ?? null;
@@ -32,15 +33,19 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         <div class="form-group">
             <label for="nome">Nome:</label>
             <input type="text" id="nome" name="nome" value="<?php echo htmlspecialchars($mesa['nome']); ?>" required>
+        </div>
         <div class="form-group">
             <label for="descricao">Descrição:</label>
             <textarea id="descricao" name="descricao"><?php echo htmlspecialchars($mesa['descricao']); ?></textarea>
+        </div>
         <div class="form-group">
             <label for="nome_do_mestre">Mestre:</label>
             <input type="text" id="nome_do_mestre" name="nome_do_mestre" value="<?php echo htmlspecialchars($mesa['nome_do_mestre']); ?>" required>
+        </div>
         <div class="form-group">
             <label for="max_capacity">Número Máximo de Jogadores:</label>
             <input type="number" id="max_capacity" name="max_capacity" value="<?php echo htmlspecialchars($mesa['max_capacity']); ?>" required>
+        </div>
         <div class="form-group">
             <label for="categoria">Categoria:</label>
             <input type="text" id="categoria" name="categoria" value="<?php echo htmlspecialchars($mesa['categoria']); ?>" required>
@@ -51,8 +56,16 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                 $users = $mesaController->getMesaParticipants($mesa_id);
                 foreach ($users as $user) {
                     echo htmlspecialchars($user['username']); 
+                    if ($user['username'] === $mesa['nome_do_mestre']) {
+                        ?>
+                        <a class="btn btn-disabled" href="#" onclick="return false;">Não é possível retirar o Mestre</a>
+                        <?php
+                    } else {
+                        ?>
+                        <a class="btn-leave" href="delete-user.php?mesa_id=<?php echo htmlspecialchars($mesa['id']); ?>&user_id=<?php echo htmlspecialchars($user['id']); ?>">Tirar jogador</a>
+                        <?php
+                    }
                     ?>
-                    <a class="btn-leave" href="delete-user.php?mesa_id=<?php echo htmlspecialchars($mesa['id']); ?>&user_id=<?php echo htmlspecialchars($user['id']); ?>">Tirar jogador</a>
                     <br>
                     <br>
                     <?php
